Add functions for retrieving/handling all available configurations
The function `get_available_configurations` returns a list of all
configurations which are available. (For a multi-config generator these
are the same as stored in `CMAKE_CONFIGURATION_TYPES`, for a
single-config generator the single result is the same as stored in
`CMAKE_BUILD_TYPE`.)

The function `get_names_with_configurations` takes some strings and
returns duplicates of them which are "configuration-dependent". That
means for each configuration there is a duplicate string returned with
that configuration appended (in uppercase letters and connected by an
underscore).

Signed-off-by: Deniz Bahadir <dbahadir@benocs.com>