Add CMake module for `eval_genex` function
This is a function that tries to evaluate generator-expressions during
the "configure" step of CMake.
Because this has to spawn an additional CMake process this is not the
fastest function.

Note:
You should test this function thoroughly for your specific generator-
expression, because it might get evaluated wrong if it is too complex.

Signed-off-by: Deniz Bahadir <dbahadir@benocs.com>