Add scripts for pre/post actions of `project` command
These scripts need to be called (respectively) before/after a call to
CMake's `project` command.

The "pre" script loads some settings required for calling the `project`
command properly. The "post" script additionally does some extra actions
after the `project` command was processed.
For example, the "post" script tries to obtain the Git revision of the
directory the currently being processed CMakeLists.txt file is located
in and stores it in a variable for later use.

Signed-off-by: Deniz Bahadir <deniz@code.bahadir.email>