Add templates for build-configuration of a project
This file is the template for the C++ source file that should be
generated by CMake (with the help of `ex_configure_file`) for each
project.

The generated source file creates a sub-namespace `module_info` in the
project's namespace containing variable constants with settings of the
associated project. Some of these settings are:
- project name
- project version (with individual components)
- build number
- configuration type (Debug, Release ...)
- ...

Signed-off-by: Deniz Bahadir <deniz@code.bahadir.email>