Add function `make_extra_project_variables`
This is a function which generates additional project-specific variables
from a project's hierarchy.

It should take `${PROJECT_NAME}` as argument, which is supposed to
represent the current project's hierarchy with each component separated
by `.` However, it is flexible enough to take some other formatted
hierarchy string and an additional separator string and derive the
project-specific variables from it.

In general, it should be called from `common-project-post-actions.cmake`
to automatically generate these variables for each project.

Signed-off-by: Deniz Bahadir <deniz@code.bahadir.email>