Add function `add_project_test`
This is a function which creates a new test-executable target and
registers it as a (unit-)test. This is an extended version of a combined
`add_executable` and `add_test` command.

The purpose is to automatically make some common configurations when
creating common unit-tests of a project and thereby reduce boiler plate.

Signed-off-by: Deniz Bahadir <deniz@code.bahadir.email>