Add function `load_project_sources`
The convenience function `load_project_sources` automatically reads in
some files which contain the list of public and private source files and
returns these to the caller's scope. (Those could then be added as
sources to the target given as argument.)

The files which contain the list of public and private source files need
to be located in directory `${PROJECT_SOURCE_DIR}/luchs/` and must be
named `project-sources_-_<name>_-_private.cmake` and
`project-sources_-_<name>_-_public.cmake`.
Thereby, `<name>` represents the given target's name unless it has the
form `${PROJECT_NAME}-<basename>` in which case `<name>` represents the
`<basename>` part

Signed-off-by: Deniz Bahadir <deniz@code.bahadir.email>