Add template for export-macros header
In order to properly support exporting symbols from shared libraries or
DLLs if hidden symbol-visibility is enabled (as is the default on
Windows) one needs to annotate classes, functions etc.
This is normally done with some export-macros, which are specific for a
project.

CMake will make sure to generate a header containing these macros for
every project from the template file introduced with this commit.

Signed-off-by: Deniz Bahadir <deniz@code.bahadir.email>