Load product-specific information as "pre" action
The file that contains product-specific information will be loaded once
from the first "pre" actions script.
By default, it needs to be provided by the project that will be built.
In case that none will be provided such a file will be generated and
that one will be loaded instead.

Signed-off-by: Deniz Bahadir <deniz@code.bahadir.email>