Add function `register_project_targets`
This function registers the given targets with the corresponding
install-components of the current project.

After some targets have been registered by that function the functions
`install_project_targets` and `install_project_debugsymbols` can now
also be called without any targets as arguments. In that case the
already registered targets will be used instead.

If `install_project_targets` is called with targets as arguments these
will automatically be registered as well.

Signed-off-by: Deniz Bahadir <deniz@code.bahadir.email>