Add function `generate_project_script`
This function can generate two scripts, the package-config script and
the dependency-loader script for the current project.

The package-config script contains the information to properly import
the current project's exported targets via CMake's `find_package`
command. It needs to be installed with the current project's "RUNTIME"
install-component.

When the current project's exported targets will be imported again via
CMake's `find_package` command it is the dependency-loader script's
responsibility to automatically try to load/import the dependencies as
well, as long as those have the same root-namespace.

Signed-off-by: Deniz Bahadir <deniz@code.bahadir.email>