Add template for dependency-loader script of a project's import-script
This file is the template for a CMake script that should be generated
and installed together with export-sets of a project. It will be called
automatically by the corresponding package-config file of the project
when loaded by CMake's `find_package` command.

When loaded, it takes care to import the targets from an export-set and
then tries to find dependencies of these newly imported targets, as long
as the name of such a dependency has the same root-namespace as the
imported targets.

Note: For common dependencies which have a different root-namespace a
project's import-script is used directly instead.

Signed-off-by: Deniz Bahadir <deniz@code.bahadir.email>