Add template for an import-script of a project
This file is the template for a CMake script that should be generated
and installed together with export-sets of a project. It will be called
automatically by the corresponding package-config file of the project
when loaded by CMake's `find_package` command.

When loaded, it makes sure to load all the individual export-sets and
then tries to find common dependencies of the newly imported targets, as
long as the name of such a dependency has a different root-namespace
than the imported targets.

Note: This can only succeed if the CMake project which calls the
`find_package` command also uses the "luchs" CMake support framework and
therefore has the `load_common_dependency` function.
Additionally, the searched dependency must be one of the "common"
dependencies for which `load_common_dependency` has explicit support.

Note: For dependencies which have the same root-namespace the
"dependency-loader" script is used instead.

Signed-off-by: Deniz Bahadir <deniz@code.bahadir.email>