Add tool.-settings function `enable_separate_debugsymbols`
This is a function which enables the generation of separate files for
debug-symbols when building.

Additionally, a helper-function `create_separate_debugsymbols_file` is
introduced which registers a post-build action for creating such
separate debug-symbols files for individual build-targets.

Note: That associated helper-function is only required to be called on
the Linux platform for build-targets that shall create separate
debug-symbols files. However, it is still recommended to also call it on
the Windows platform (although it effectively is a no-op) because it
does not harm and allows the target to be built uniformly on different
platforms without the need to change any `CMakeLists.txt` file.

Signed-off-by: Deniz Bahadir <deniz@code.bahadir.email>