Add tool.-settings function `generate_entry_property_file`
This is a function which generates the "Directory.Build.props" entry
property file (from the template also introduced by this commit) that
will automatically be loaded when building with MSBuild / Visual Studio.
It will be generated at the root of the build directory and when loaded
will automatically load other (language-specific) property files from
the `luchs.MSBuild.Props` subdirectory (located in the same build
directory).

Note: Those other language-specific property files will be introduced
with the following commits.

Signed-off-by: Deniz Bahadir <deniz@code.bahadir.email>