Add tool.-settings function `generate_cxx_property_file`
This is a function which generates the "General.Cpp.props" property file
(from the template also introduced by this commit) that will
automatically be loaded (indirectly through "Directory.Build.props")
when building C++ code with MSBuild / Visual Studio.
It will be generated in the subdirectory `luchs.MSBuild.Props` of the
root of the build directory.

Note: Currently, the generated property file does not contain any
specific settings, but this might change with some further commits.

Signed-off-by: Deniz Bahadir <deniz@code.bahadir.email>