Add tool.-settings function `generate_csharp_property_file`
This is a function which generates the "General.CSharp.props" property
file that will automatically be loaded (indirectly through
"Directory.Build.props") when building C# code with MSBuild / Visual
Studio.
It will be generated in the subdirectory `luchs.MSBuild.Props` of the
root of the build directory.

Note: Currently, no template exists for this file so that it will not be
generated. A template file might get introduced later if more support
for C# will be added.

Signed-off-by: Deniz Bahadir <deniz@code.bahadir.email>