Add tool.-settings function `remove_fullpath_to_pdb_from_dll_and_exe`
This is a function which instructs the linker of MSVC (or Clang-cl) to
only store the associated PDB file's name in the built executable/DLL
instead of storing the full path.

In order for this to work properly with any Visual Studio generator one
needs to properly adjust the generated property file.
Note: This adjustment will be done with the following commit.

Signed-off-by: Deniz Bahadir <deniz@code.bahadir.email>