Use file-sets for public sources from within `add_project_*` functions
If a CMake versions is used that supports file-sets, the convenience
functions `add_project_test`, `add_project_library` and
`add_project_executable` now use this feature. The list of public
sources (aka headers) obtained with the help of `load_project_sources`
will then be added to the `HEADERS` file-set of the specific target.

Signed-off-by: Deniz Bahadir <deniz@code.bahadir.email>