Extend function `add_project_test` to support test-frameworks
This function now accepts an option `NO_SEPARATE_TEST_RUNS` which
instructs CTest to run the test-executable as a whole, instead of
running each contained test individually (which is the default for
test-frameworks).

Furthermore, the new option `TEST_FRAMEWORK` determines what
test-framework to use for the new test (and therefore against which
test-framework a dependency should be declared). If that option is not
given a default value is taken from the variable
`LUCHS_DEFAULT_TEST_FRAMEWORK` (if it is set).

The value `NONE` means that no test-framework is used and therefore no
extra steps will be required.
Any other value, though, depicts the name of the test-framework to use.

Support for those other test-frameworks can be simply added by providing
two properly named files in the modules subdirectory `TestFrameworks`
that implement the required steps:
* File `GetLinkTargets_<test-framework>.cmake` needs to set a variable
  `TEST_FRAMEWORK_LINK_TARGETS` which contains the list of targets that
  need to be linked (via `target_link_libraries`) by the test target.
* File `AddDiscoverableTests_<test-framework>.cmake` on the other hand
  needs to register the test target in such a way that its individual
  tests can be run separately. In best case, the invidivual tests are
  discoverable (at runtime) from the test executable.
  (For example, CMake's `GoogleTest` module provides such a function
  `gtest_discover_tests` which is able to do this.)

Note, however, that it is the user's responsibility to load the
test-framework (preferably via using function `load_common_dependency`)
before it is used by `add_project_test`. Otherwise, an error is very
likely.

Signed-off-by: Deniz Bahadir <deniz@code.bahadir.email>