Add template for an import-script of a CMake group-package
This file is the template for a CMake script that should be generated
and installed for a project that functions as the root/entry-point to
other sub-projects. It will be called automatically by the corresponding
package-config file of the project when its package will be loaded by
CMake's `find_package` command.

It takes care of loading individual (bundled) components of the
associated package. Those individual components most likely correspond
to other (sub-)projects of the current (root-)project. In general, this
should be visible by the (sub-)projects' names, if they use the
group-project's name as their root-namespace.

A call to `find_package` could then look like this:
```
find_package( foo COMPONENT bar baz::a baz::b )
```
It would then result in loading the import-script generated from the
script template introduced by this commit for the "foo"-project. That
would then try to load the components (aka projects) `foo::bar`,
`foo::baz::a` and `foo::baz::b`.

Signed-off-by: Deniz Bahadir <deniz@code.bahadir.email>