Add template of an install script for using meta install-components
This file is the template for a CMake script that should automatically
be processed when installing build-artifacts using any of the meta
install-components ("RUNTIME", "DEVELOPMENT", "PLUGINS" or
"DEBUGSYMBOLS").

It will grep the general install-scripts for the names of specific
install-components that are associated with the meta install-component
currently used and triggers installing these specific
install-components.

The meta install-components are therefore a convenient way to install
all install-components associated with it.

Signed-off-by: Deniz Bahadir <deniz@code.bahadir.email>