Add template for a make-global script of a project's import-script
This file is the template for a CMake script that should be generated
and installed together with export-sets of a project. It will be called
automatically by the corresponding package-config file of the project
when loaded by CMake's `find_package` command.

When loaded, it takes care to promote the imported targets from an
export-set to global scope if the corresponding
`<package-name>_IMPORT_GLOBALLY` variable is set to `TRUE`.

Signed-off-by: Deniz Bahadir <deniz@code.bahadir.email>