Add template for project-specific CPack config
The path to the file generated from this template must be stored as
CPack variable `CPACK_OUTPUT_CONFIG_FILE` in the configuration settings
for CPack. The file will then be loaded automatically by CPack for each
individual CPack-generator. It will delegate to a CPack-generator
specific script which will do further configuring for packaging with
that CPack-generator.

Signed-off-by: Deniz Bahadir <deniz@code.bahadir.email>