Add template for CMakeLists.txt file for packaging a project
At the end of each project's `CMakeLists.txt` file the settings for
packaging that project (using CPack) should be made.
The current template can be used to generate a `CMakeLists.txt` file
(preferably in a subdirectory `packaging`) which then should be loaded
(via CMake's `add_subdirectory` command) at the end of each project's
`CMakeLists.txt` file and that makes the settings for packaging.

Signed-off-by: Deniz Bahadir <deniz@code.bahadir.email>