Add function `generate_cpack_deb_extra_settings_file`
This function needs to be called when preparing CPack variables for
generating DEB-packages. It will generate an additional file that
contains extra CPack settings for generating DEB-packages and which
therefore must be loaded by CPack when generating DEB-packages.

Note, however, that currently the generated file does not contain any
additional information which is useful for CPack. (Such information will
be added with the next commits.)

Signed-off-by: Deniz Bahadir <deniz@code.bahadir.email>