Add tool.-settings function `enable_unicode_charset_globally`
This is a function which defines the preprocessor macros `UNICODE` and
`_UNICODE`, which will also be considered by CMake when generating
MSBuild project files, in order to determine if the Unicode or the
Multi-Byte character-set should be used (on Windows).

When using the Visual Studio generator, calling this function will
generate MSBuild project files with the Unicode character-set. However,
if this is not desired for a specific MSBuild project file, then at
least the `_UNICODE` macro definition must be undefined again for the
corresponding `CMakeLists.txt` file!

Signed-off-by: Deniz Bahadir <deniz@code.bahadir.email>