Add template for a common, project-specific MSBuild property file
This file is the template for a MSBuild property file which shall be
loaded by CMake targets of a project. By setting a target's
`VS_USER_PROPS` property to the path of the MSBuild property file
generated from this template it will be loaded automatically for that
target.

Because the `VS_USER_PROPS` CMake property (ab)uses the "user property"
mechanism to load the file generated from this template, the first thing
that it does is to restore the original "user property" mechanism.

Additionally, it then tries to load a common, global MSBuild property
file (if it exists) and afterwards a local MSBuild property file
specific for the CMake target it is used for.

Signed-off-by: Deniz Bahadir <deniz@code.bahadir.email>