Add function `generate_extra_project_msbuild_property_file`
This is a function which generates a project-specific MSBuild property
file.

In general, it should be called as "post" action of a project in order
to automatically generate the MSBuild property file for each project.
Of course, this only is needed when using the Visual Studio generator
(therefore effectively using MSBuild as build-tool).

Signed-off-by: Deniz Bahadir <deniz@code.bahadir.email>