Add template for a common, global MSBuild property file
This file is the template for a common, global MSBuild property file
which shall be loaded by CMake targets of a project. It will be loaded
automatically through the common, project-specific property file of each
target.

It contains (global) settings that were provided by 'luchs' and that do
not have another way to be set.

Note: Right now this file does not contain any specific settings. They
need to be added with further commits.

Signed-off-by: Deniz Bahadir <deniz@code.bahadir.email>